+++ src/main/java/com/example/service/CompanyDeleteService.java
@@ -1,7 +1,7 @@
 package com.example.service;
 
 public interface CompanyDeleteService {
-	void deleteById(Long id);
+	void deleteById(int id);
 
 
 }
