+++ src/main/java/com/example/service/CompanyFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Company;
 
 public interface CompanyFetchService {
-	Company fetchById(Long id);
+	Company fetchById(int id);
 
 }
