+++ src/main/java/com/example/service/DepartmentFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Department;
 
 public interface DepartmentFetchService {
-	Department fetchById(Long id);
+	Department fetchById(int id);
 
 }
